bonusy prcanq??

+++ libft/ft_atol.c
@@ -13,6 +13,8 @@
 #include "libft.h"
 #include <stdio.h>
 
+//# ifeq ($(shell uname -s), MINGW64_NT-6.1-7601)
+
 void	err_push_for_lib(int a)
 {
 	if (a)
@@ -25,7 +27,7 @@ void	err_push_for_lib(int a)
 long	ft_atol(char *str)
 {
 	int		sign;
-	long	num;
+	long long	num;
 
 	num = 0;
 	sign = 1;
